Adding CRUD for Revenu and Depenses

+++ database/migrations/2021_09_19_154155_create_depenses_table.php
@@ -18,7 +18,7 @@ class CreateDepensesTable extends Migration
             $table->integer('id_user');
             $table->string('nom');
             $table->text('description')->nullable();
-            $table->float('montant');
+            $table->double('montant');
             $table->timestamps();
         });
     }
