+++ database/migrations/2021_09_19_154631_create_mouvements_table.php
@@ -20,7 +20,7 @@ class CreateMouvementsTable extends Migration
             $table->id();
             $table->integer('id_user');
             $table->integer('source');
-            $table->float('montant');
+            $table->double('montant');
             $table->timestamps();
         });
     }
